Create code.dart
See, it can be run using the android studio and images can be changed as per the need. It demonstrate how the basic outline of any app is done using the flutter and can be improved in the nearby future. The concept of One hot encoder is also quite useful, when working with app development as there is no time lag in the development procedure and changes made are visible instantaneously.